backend/requirements.txt
flask==2.3.3
flask-cors==4.0.0
opencv-python==4.8.1.78
torch==2.0.1
torchvision==0.15.2
ultralytics==8.0.196
easyocr==1.7.0
Pillow==10.0.1
numpy==1.24.3
gtts==2.3.2
requests==2.31.0